Move Decimal9 and Uint128 into math package
